+++ nama/__init__.py
@@ -1,4 +1,5 @@
-from .matcher import Matcher
+from .match_groups import MatchGroups, read_csv, from_df
 from .utils import *
-from .models.similarity_model import SimilarityModel
+from .models.similarity_model import SimilarityModel, load_similarity_model
+from .models.embeddings import load_embeddings
 from .scoring import score_predicted
