+++ nama/utils.py
@@ -1,30 +1,5 @@
-# import nltk
-# from nltk.corpus import stopwords
-# from pathlib import Path
-import numpy as np
-import pandas as pd
-# from matplotlib import pyplot as plt
-# from tqdm import tqdm
-# import matplotlib.pyplot as plt
-# import sys
-import torch
-# import argparse
-# from datetime import datetime
 import re
-# import os
-# from pathlib import Path
-from collections import Counter
-# from itertools import islice
-import pandas as pd
-import numpy as np
-# import networkx as nx
-# import matplotlib.pyplot as plt
-# import matplotlib as mplt
-from zipfile import ZipFile
-import pickle
-# from .scoring import score_predicted, split_on_groups
-from .matcher import Matcher
-from .models.embeddings import Embeddings
+
 
 # Preprocessors
 
@@ -81,144 +56,8 @@ def simplify_corp(s):
 
     return s
 
-# Tokenizers
 
-def ngrams(string, n=2):
-    """
-    Generates all n-grams of the given string.
-
-    Parameters
-    ----------
-    string: str
-        The string from which the n-grams will be generated.
-    n: int, optional
-        The number of characters in each n-gram. Default is 2.
-
-    Yields
-    ------
-    str
-        An n-gram of the string.
-
-    """
-    for i in range(len(string) - n + 1):
-        yield string[i:i + n]
-
-
-def nmgrams(string, n=1, m=3):
-    """
-    Generates all n-grams of the given string where the n-grams are of length between n and m.
-
-    Parameters
-    ----------
-    string: str
-        The string from which the n-grams will be generated.
-    n: int, optional
-        The minimum length of the n-grams to be generated. Default is 1.
-    m: int, optional
-        The maximum length of the n-grams to be generated. Default is 3.
-
-    Yields
-    ------
-    str
-        An n-gram of the string where the length is between n and m.
-
-    """
-    for j in range(n, m + 1):
-        for i in range(len(string) - j + 1):
-            yield string[i:i + j]
-
-
-def words(string):
-    """
-    Generates all words in the given string.
-
-    Parameters
-    ----------
-    string: str
-        The string from which the words will be extracted.
-
-    Yields
-    ------
-    str
-        A word extracted from the string.
-
-    """
-    for m in re.finditer(r'[A-Za-z0-9]+', string):
-        yield m.group(0)
-
-# Similarity Measures
-
-def jaccard_similarity(set0, set1, weights):
-    """
-    Calculates the Jaccard similarity between two sets of terms.
-
-    Parameters
-    ----------
-    set0: set
-        The first set of terms.
-    set1: set
-        The second set of terms.
-    weights: dict
-        A dictionary mapping terms to weights.
-
-    Returns
-    -------
-    float
-        The Jaccard similarity between set0 and set1.
-
-    """
-    intersection = set0 & set1
-
-    if not intersection:
-        return 0
-
-    union = set0 | set1
-    denominator = sum(weights[t] for t in union)
-
-    if not denominator:
-        return 0
-
-    numerator = sum(weights[t] for t in intersection)
-    return numerator / denominator
-
-
-def cosine_similarity(set0, set1, weights):
-    """
-    Calculates the cosine similarity between two sets of terms.
-
-    Parameters
-    ----------
-    set0: set
-        The first set of terms.
-    set1: set
-        The second set of terms.
-    weights: dict
-        A dictionary mapping terms to weights.
-
-    Returns
-    -------
-    float
-        The cosine similarity between set0 and set1.
-
-    """
-    intersection = set0 & set1
-
-    if not intersection:
-        return 0
-
-    length0 = np.sqrt(sum(weights[t]**2 for t in set0))
-    length1 = np.sqrt(sum(weights[t]**2 for t in set1))
-
-    denominator = length0 * length1
-    if not denominator:
-        return 0
-
-    numerator = sum(weights[t]**2 for t in intersection)
-
-    return numerator / denominator
-
-
-# Matcher Tools
+# MatchGroups Tools
 
 # def plot(matchers, strings, matcher_names=None, ax=None):
 #     """
@@ -227,7 +66,7 @@ def cosine_similarity(set0, set1, weights):
 
 #     Parameters
 #     ----------
-#     matchers : Matcher or list of Matcher
+#     matchers : MatchGroups or list of MatchGroups
 #         a matcher or list of matchers to plot
 #     strings : str or list of str
 #         a string or list of strings to plot (all connected strings will also be plotted)
@@ -242,7 +81,7 @@ def cosine_similarity(set0, set1, weights):
 #         The matplotlib axis object with the plot.
 #     """
 
-#     if isinstance(matchers, Matcher):
+#     if isinstance(matchers, MatchGroups):
 #         matchers = [matchers]
 
 #     if isinstance(strings, str):
@@ -334,193 +173,3 @@ def cosine_similarity(set0, set1, weights):
 #                ncol=3, borderaxespad=0)
 
 #     return ax
-
-
-def from_df(
-        df,
-        match_format='detect',
-        pair_columns=[
-            'string0',
-            'string1'],
-    string_column='string',
-    group_column='group',
-        count_column='count'):
-    """
-    Construct a new matcher from a pandas DataFrame.
-
-    Parameters
-    ----------
-    df : pandas.DataFrame
-        The input dataframe.
-    match_format : str, optional
-        The format of the dataframe, by default "detect".
-        It can be one of ['unmatched', 'groups', 'pairs', 'detect'].
-    pair_columns : list of str, optional
-        The columns names containing the string pairs, by default ['string0','string1'].
-    string_column : str, optional
-        The column name containing the strings, by default 'string'.
-    group_column : str, optional
-        The column name containing the groups, by default 'group'.
-    count_column : str, optional
-        The column name containing the counts, by default 'count'.
-
-    Returns
-    -------
-    Matcher
-        The constructed Matcher object.
-
-    Raises
-    ------
-    ValueError
-        If the input `match_format` is not one of ['unmatched', 'groups', 'pairs', 'detect'].
-    ValueError
-        If the `match_format` is 'detect' and the input dataframe format could not be inferred.
-
-    Notes
-    -----
-    The function accepts two formats of the input dataframe:
-
-        - "groups": The standard format for a matcher dataframe. It includes a
-          string column, and a "group" column that contains group labels, and an
-          optional "count" column. These three columns completely describe a
-          matcher object, allowing lossless matcher -> dataframe -> matcher
-          conversion (though the specific group labels in the dataframe will be
-          ignored and rebuilt in the new matcher).
-
-        - "pairs": The dataframe includes two string columns, and each row indicates
-          a link between a pair of strings. A new matcher will be constructed by
-          uniting each pair of strings.
-    """
-
-    if match_format not in ['unmatched', 'groups', 'pairs', 'detect']:
-        raise ValueError(
-            'match_format must be one of "unmatched", "groups", "pairs", or "detect"')
-
-    # Create an empty matcher
-    matcher = Matcher()
-
-    if match_format == 'detect':
-        if (string_column in df.columns):
-            if group_column is None:
-                match_format = 'unmatched'
-            elif (group_column in df.columns):
-                match_format = 'groups'
-        elif set(df.columns) == set(pair_columns):
-            match_format = 'pairs'
-
-    if match_format == 'detect':
-        raise ValueError('Could not infer valid dataframe format from input')
-
-    if count_column in df.columns:
-        counts = df[count_column].values
-    else:
-        counts = np.ones(len(df))
-
-    if match_format == 'unmatched':
-        strings = df[string_column].values
-
-        # Build the matcher
-        matcher.counts = Counter({s: int(c) for s, c in zip(strings, counts)})
-        matcher.labels = {s: s for s in strings}
-        matcher.groups = {s: [s] for s in strings}
-
-    elif match_format == 'groups':
-
-        strings = df[string_column].values
-        group_ids = df[group_column].values
-
-        # Sort by group and string count
-        g_sort = np.lexsort((counts, group_ids))
-        group_ids = group_ids[g_sort]
-        strings = strings[g_sort]
-        counts = counts[g_sort]
-
-        # Identify group boundaries and split locations
-        split_locs = np.nonzero(group_ids[1:] != group_ids[:-1])[0] + 1
-
-        # Get grouped strings as separate arrays
-        groups = np.split(strings, split_locs)
-
-        # Build the matcher
-        matcher.counts = Counter({s: int(c) for s, c in zip(strings, counts)})
-        matcher.labels = {s: g[-1] for g in groups for s in g}
-        matcher.groups = {g[-1]: list(g) for g in groups}
-
-    elif match_format == 'pairs':
-        # TODO: Allow pairs data to use counts
-        for pair_column in pair_columns:
-            matcher.add_strings(df[pair_column].values, inplace=True)
-
-        # There are several ways to unite pairs
-        # Guessing it is most efficient to "group by" one of the string columns
-        groups = {s: pair[1] for pair in df[pair_columns].values for s in pair}
-
-        matcher.unite(groups, inplace=True)
-
-    return matcher
-
-
-def read_csv(
-        filename,
-        match_format='detect',
-        pair_columns=[
-            'string0',
-            'string1'],
-    string_column='string',
-    group_column='group',
-    count_column='count',
-        **pandas_args):
-    """
-    Read a csv file and construct a new matcher.
-
-    Parameters
-    ----------
-    filename : str
-        The path to the csv file.
-    match_format : str, optional (default='detect')
-        One of "unmatched", "groups", "pairs", or "detect".
-    pair_columns : list of str, optional (default=['string0', 'string1'])
-        Two string columns to use if match_format='pairs'.
-    string_column : str, optional (default='string')
-        Column name for string values in match_format='unmatched' or 'groups'.
-    group_column : str, optional (default='group')
-        Column name for group values in match_format='groups'.
-    count_column : str, optional (default='count')
-        Column name for count values in match_format='unmatched' or 'groups'.
-    **pandas_args : optional
-        Optional arguments to pass to `pandas.read_csv`.
-
-    Returns
-    -------
-    Matcher
-        A new matcher built from the csv file.
-    """
-    df = pd.read_csv(filename, **pandas_args, na_filter=False)
-    df = df.astype(str)
-
-    return from_df(df, match_format=match_format, pair_columns=pair_columns,
-                   string_column=string_column, group_column=group_column,
-                   count_column=count_column)
-
-
-def load_embeddings(f):
-    """
-    Load embeddings from custom zipped archive format
-    """
-    with ZipFile(f,'r') as zip:
-        score_model = pickle.loads(zip.read('score_model.pkl'))
-        weighting_function = pickle.loads(zip.read('weighting_function.pkl'))
-        strings_df = pd.read_csv(zip.open('strings.csv'),na_filter=False)
-        V = np.load(zip.open('V.npy'))
-
-        return Embeddings(
-                            strings=strings_df['string'].values,
-                            counts=torch.tensor(strings_df['count'].values),
-                            score_model=score_model,
-                            weighting_function=weighting_function,
-                            V=torch.tensor(V)
-                            )
-
-
-def load_similarity_model(f,map_location='cpu',**kwargs):
-    return torch.load(f,map_location=map_location,**kwargs)
